refactor: Rename Binding Implementation Modules By Domain
