+++ rust/python-bindings/src/schema_store.rs
